提交更新，增加websocket服务(netty)...

+++ ctm_project/api/src/main/java/eqlee/ctm/api/pay/controller/PayController.java
@@ -22,10 +22,10 @@ import eqlee.ctm.api.pay.entity.PayToken;
 import eqlee.ctm.api.pay.entity.query.ResultQuery;
 import eqlee.ctm.api.pay.entity.vo.PayTokenVo;
 import eqlee.ctm.api.pay.entity.vo.ResultTokenVo;
-import eqlee.ctm.api.pay.entity.vo.TokenVo;
 import eqlee.ctm.api.pay.entity.vo.sucFailVo;
 import eqlee.ctm.api.pay.service.IPayService;
 import eqlee.ctm.api.pay.vilidata.PayData;
+import eqlee.ctm.api.pay.vilidata.PayTokenUtils;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
 import io.swagger.annotations.ApiImplicitParams;
@@ -60,36 +60,12 @@ public class PayController {
     @Autowired
     private LocalUser localUser;
 
+    @Autowired
+    private PayTokenUtils payTokenUtils;
+
     IdGenerator idGenerator = new IdGenerator();
 
     private final String URL = "http://pay.wapi.eqlee.com";
-
-    private final String AccessKey = "157222853200009";
-
-    private final String AccessKeySecret = "74c172912d9b4631bc72e567c27b2ece";
-
-    /**
-     * 获取token
-     * @return
-     * @throws Exception
-     */
-    public Object getPayToken () throws Exception{
-
-        String url = URL +"/v1/Token/Token";
-
-        PayToken payToken = new PayToken();
-        payToken.setAccessKey(AccessKey);
-        payToken.setAccessKeySecret(AccessKeySecret);
-
-        String s = JSONObject.toJSONString(payToken);
-        HttpResult httpResult = apiService.doPost(url, s);
-
-        PayTokenVo vo = JSONObject.parseObject(httpResult.getBody(), PayTokenVo.class);
-
-        ResultTokenVo result = vo.getResult();
-        return result.getToken();
-    }
-
 
 
     @ApiOperation(value = "微信扫码支付", notes = "微信扫码支付")
@@ -108,7 +84,7 @@ public class PayController {
         String url = URL + "/v1/WeChatPay/GetQRCode?payOrderSerialNumber=" + payOrderSerialNumber +"&Money=" +Money
                 + "&productName=" + productName + "&callbackUrl=" +callbackUrl;
 
-        Object token = getPayToken();
+        String token = payTokenUtils.getMapToken();
         String tokenString = "Bearer " +token;
 
         HttpResult httpResult = apiService.get(url, tokenString);
@@ -121,7 +97,7 @@ public class PayController {
 
 
     /**
-     * 支付完成调用此接口
+     * 微信支付完成调用此接口
      * @param vo
      * @return
      */
@@ -137,18 +113,41 @@ public class PayController {
             pay.setPayStatu(1);
             pay.setPayDate(LocalDateTime.now());
             pay.setThirdPartyNumber(vo.getThirdPayOrderId());
+            pay.setPayType(0);
             PayResult result = payService.insertPayInfo(pay);
             //修改报名表支付状态
-            payService.updateApplyPayStatus(vo.getPayOrderSerialNumber());
+            payService.updateApplyPayStatus(vo.getPayOrderSerialNumber(),0);
             return Result.SUCCESS(result);
         }
         //支付失败
-        pay.setId(idGenerator.getNumberId());
-        pay.setApplyNo(vo.getPayOrderSerialNumber());
-        pay.setPayStatu(2);
-        pay.setPayDate(LocalDateTime.now());
-        pay.setThirdPartyNumber(vo.getThirdPayOrderId());
-        payService.insertPayInfo(pay);
+        PayResult applyInfo = payService.deleteApplyInfo(vo.getPayOrderSerialNumber());
+        return Result.SUCCESS(applyInfo);
+    }
+
+    /**
+     * 支付宝支付完成调用此接口
+     * @param vo
+     * @return
+     */
+    @PostMapping("/sucAliFail")
+    @CrossOrigin
+    @IgnoreResponseAdvice
+    public Result sucAliFail (@RequestBody sucFailVo vo) {
+        Pay pay = new Pay();
+        if ("SUCCESS".equals(vo.getMessage())) {
+            //支付成功
+            pay.setId(idGenerator.getNumberId());
+            pay.setApplyNo(vo.getPayOrderSerialNumber());
+            pay.setPayStatu(1);
+            pay.setPayDate(LocalDateTime.now());
+            pay.setThirdPartyNumber(vo.getThirdPayOrderId());
+            pay.setPayType(1);
+            PayResult result = payService.insertPayInfo(pay);
+            //修改报名表支付状态
+            payService.updateApplyPayStatus(vo.getPayOrderSerialNumber(),1);
+            return Result.SUCCESS(result);
+        }
+        //支付失败
         PayResult applyInfo = payService.deleteApplyInfo(vo.getPayOrderSerialNumber());
         return Result.SUCCESS(applyInfo);
     }
@@ -214,7 +213,7 @@ public class PayController {
                 + "&productName=" + productName + "&callbackUrl=" +callbackUrl + "&openId=" +openId;
 
         //获取token
-        Object token = getPayToken();
+        String token = payTokenUtils.getMapToken();
         String tokenString = "Bearer " +token;
 
         HttpResult httpResult = apiService.get(url, tokenString);
@@ -228,15 +227,15 @@ public class PayController {
     @ApiOperation(value = "支付完成查询接口", notes = "支付完成查询接口")
     @ApiImplicitParam(name = "applyNo", value = "订单号", required = true, dataType = "String", paramType = "path")
     @GetMapping("/queryPayResult")
-    @CrossOrigin
     @IgnoreResponseAdvice
+    @CrossOrigin
     public Object queryPayResult(@RequestParam("applyNo") String applyNo) throws Exception{
 
         //查询支付结果
         String url = URL + "/v1/WeChatPay/QueryOrderState?payOrderSerialNumber=" + applyNo;
 
         //获取token
-        Object token = getPayToken();
+        String token = payTokenUtils.getMapToken();
         String tokenString = "Bearer " +token;
 
         HttpResult httpResult = apiService.get(url, tokenString);
@@ -270,7 +269,7 @@ public class PayController {
                 + "&refundFee=" +refundFee;
 
         //获取token
-        Object token = getPayToken();
+        String token = payTokenUtils.getMapToken();
         String tokenString = "Bearer " +token;
 
         HttpResult httpResult = apiService.get(url, tokenString);
